requirements.txt
requests==2.22.0
discord.py==1.3.3
discord==1.0.1
python-dotenv==0.13.0
